Move PublishOptions methods to Publish
